Move TechnoVEConfigEntry to init file
