+++ homeassistant/components/technove/binary_sensor.py
@@ -16,9 +16,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from . import TechnoVEConfigEntry
 from .coordinator import TechnoVEDataUpdateCoordinator
 from .entity import TechnoVEEntity
-from .typing import TechnoVEConfigEntry
 
 
 @dataclass(frozen=True, kw_only=True)
