+++ homeassistant/components/technove/coordinator.py
@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from technove import Station as TechnoVEStation, TechnoVE, TechnoVEError
 
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
@@ -12,11 +13,14 @@ from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, Upda
 
 from .const import DOMAIN, LOGGER, SCAN_INTERVAL
 
+if TYPE_CHECKING:
+    from . import TechnoVEConfigEntry
+
 
 class TechnoVEDataUpdateCoordinator(DataUpdateCoordinator[TechnoVEStation]):
     """Class to manage fetching TechnoVE data from single endpoint."""
 
-    def __init__(self, hass: HomeAssistant, entry: ConfigEntry) -> None:
+    def __init__(self, hass: HomeAssistant, entry: TechnoVEConfigEntry) -> None:
         """Initialize global TechnoVE data updater."""
         self.technove = TechnoVE(
             entry.data[CONF_HOST],
