+++ homeassistant/components/technove/switch.py
@@ -13,10 +13,10 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from . import TechnoVEConfigEntry
 from .coordinator import TechnoVEDataUpdateCoordinator
 from .entity import TechnoVEEntity
 from .helpers import technove_exception_handler
-from .typing import TechnoVEConfigEntry
 
 
 @dataclass(frozen=True, kw_only=True)
